+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.2.0
 
 - Renamed base method from `fileSizeToString` to `humanFileSize()`.
-- Added extension method `humanFileSize` on `num`, `Decimal`, `BigInt` and `Rational`.
+- Added extension method `humanFileSize` on `num`.
 
 ## 0.1.2